Fast API - Installation
To install packages, you can use "pip install <PACKAGE>". But in order to have a track of your dependencies,
is better to have a file that gathers all the required dependencies. In this commit, we are creating a new file called
"requirements.txt", adding the required libraries, fastapi and uvicorn (used for serving the application) are the
only one for now.

To install all the listed dependencies in this file, just run "pip install -r requirements.txt" preferably inside a virtual
environment.