Fast API - First endpoint
Adding this lines, you are importing FastAPI, instantiating it and defining a function that handles
a GET http request from root and returns "Hello World" as a response.
Here is necesary to use decorators, in this case the decorator "@app.get" gives the ability to
handle the request to the next defined function.